+++ packages/vue/src/switch/switch-context.ts
@@ -1,0 +1,4 @@
+import { createContext } from '../context'
+import type { UseSwitchReturn } from './use-switch'
+
+export const [SwitchProvider, useSwitchContext] = createContext<UseSwitchReturn>('SwitchContext')